% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{capitalize_words}
\alias{capitalize_words}
\title{Capitalizes the first letter of every word}
\usage{
capitalize_words(words, lowercase_of = TRUE)
}
\arguments{
\item{words}{A string or vector of strings with words you want capitalized}

\item{lowercase_of}{If TRUE (default), keeps the string " of " to be lowercased as is custom in
English writing (e.g. District of Columbia).}
}
\value{
The original string with the first letter of each word capitalized
}
\description{
Capitalizes the first letter of every word
}
\examples{
capitalize_words("district of columbia")
}
