% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{clean_tex_files}
\alias{clean_tex_files}
\title{Prepare .tex output from kableExtra for Overleaf}
\usage{
clean_tex_files(files = "", all_tex = FALSE)
}
\arguments{
\item{files}{A string or vector of strings naming the .tex files to use.}

\item{all_tex}{A binary variable (default is FALSE) where if TRUE applies code to every .tex
file in working directory. Overrides any value in `files`.}
}
\description{
When kableExtra makes .tex files it adds lines to create a full .tex
}
