% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_incidents.R
\name{coordinate_splitter}
\alias{coordinate_splitter}
\title{Coordinate Splitter}
\usage{
coordinate_splitter(dataset, column_name)
}
\arguments{
\item{column_name}{}
}
\description{
Coordinate Splitter
}

