% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{crime_data}
\alias{crime_data}
\title{crime}
\usage{
crime_data(dataset_name)
}
\arguments{
\item{dataset_name}{Enter the name of the dataset you want to download. Full list of
datasets are found by entering data(available_crime_data)}
}
\value{
Data frame with crime data
}
\description{
crime
}
\examples{
crime_data("bloomington_use_force")
}

