% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_incidents.R
\name{date_column_maker}
\alias{date_column_maker}
\title{Creates the date value columns}
\usage{
date_column_maker(dataset)
}
\arguments{
\item{dataset}{Crime data.table}
}
\value{
data.table
}
\description{
Creates the date value columns
}
\examples{
1 + 3
}

