% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{fix_column_names}
\alias{fix_column_names}
\title{Makes names lowercase and replaces all punctuation with underscore.}
\usage{
fix_column_names(.names)
}
\arguments{
\item{names}{A string or vector of strings.}
}
\value{
The inputted string or vector of strings now lowercased and
with punctuation replaced with a single underscore.
}
\description{
Makes names lowercase and replaces all punctuation with underscore.
}
\examples{
fix_column_names("BAD.-columnNAME")
}
