% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{indicate_outliers}
\alias{indicate_outliers}
\title{Title}
\usage{
indicate_outliers(data, select_numeric_variables = NULL, group_variable,
  outlier_std_dev_value = 1.96, zero_is_outlier = FALSE)
}
\arguments{
\item{zero_is_outlier}{}
}
\description{
Title
}
