% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_codebooks.R
\name{lemasNames}
\alias{lemasNames}
\title{Fixes column names in LEMAS datasets from 1987 to 2003}
\usage{
lemasNames(lemas_dataset, year)
}
\arguments{
\item{lemas_dataset}{The LEMAS dataset}

\item{year}{The year of the LEMAS dataset}
}
\value{
The same data.frame as inputted but with the correct column names
}
\description{
Fixes column names in LEMAS datasets from 1987 to 2003
}
\examples{
example <- data.frame(V22 = 1, V30 = 0)
lemasNames(example, 1999)

}

