% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_codebooks.R
\docType{data}
\name{LEOKA_ColNames_1994_1997}
\alias{LEOKA_ColNames_1994_1997}
\title{Column name and code for the UCR "Police Employee" (LEOKA)
 dataset from 1994-1997}
\format{A data frame with 2255 rows and 2 variables:
\describe{
  \item{column_name}{The column name for the given column, according
  to the codebook.}
  \item{column_code}{The code for the given column, according to the codebook.
  This will be the column name of the inputted dataset.}
  ...
}}
\source{
\url{http://www.icpsr.umich.edu/icpsrweb/NACJD/series/57/studies/9028}
}
\usage{
LEOKA_ColNames_1994_1997
}
\description{
A dataset containing the name and code corresponding to columns
in the Uniform Crime Report "Police Employee" (LEOKA) dataset for
years 1994-1997
}
\keyword{datasets}

