% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_codebooks.R
\docType{data}
\name{LEOKA_ColNames_1998_2014}
\alias{LEOKA_ColNames_1998_2014}
\title{Column name and code for the UCR "Police Employee" (LEOKA)
 dataset from 1998-2014}
\format{A data frame with 2263 rows and 2 variables:
\describe{
  \item{column_name}{The column name for the given column, according
  to the codebook.}
  \item{column_code}{The code for the given column, according to the codebook.
  This will be the column name of the inputted dataset.}
  ...
}}
\source{
\url{http://www.icpsr.umich.edu/icpsrweb/NACJD/studies/36395}
}
\usage{
LEOKA_ColNames_1998_2014
}
\description{
A dataset containing the name and code corresponding to columns
in the Uniform Crime Report "Police Employee" (LEOKA) dataset for
years 1998-2014
}
\keyword{datasets}

