% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_average_linegraph}
\alias{make_average_linegraph}
\title{Create a line graph with 95\% confidence interval bars}
\usage{
make_average_linegraph(
  data,
  x_col,
  y_col,
  confident_interval_error_bars = FALSE
)
}
\arguments{
\item{confident_interval_error_bars}{}
}
\value{

}
\description{
Create a line graph with 95\% confidence interval bars
}
\examples{
data = data.frame(x = sample(15:25, size = 200, replace = TRUE),
y = sample(1:100, size = 200, replace = TRUE))
make_average_linegraph(data, "x", "y")
make_average_linegraph(data, "x", "y", confident_interval_error_bars = TRUE)
}
