% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_barplots}
\alias{make_barplots}
\title{Make a nice-looking barplot.}
\usage{
make_barplots(data, column, count = TRUE, title = NULL, ylab = NULL)
}
\arguments{
\item{data}{A data.frame with the data you want to graph.}

\item{column}{A string with the name of the column you want to make the plot from.}

\item{count}{A boolean (default TRUE) indicating if you want the barplot to show a count
of the column values or a percent.}

\item{title}{A string with the text you want as the title.}

\item{ylab}{A string with the text you want as the y-axis label.}
}
\value{
A barplot object.
}
\description{
Make a nice-looking barplot.
}
\examples{
make_barplots(mtcars, "cyl")

make_barplots(mtcars, "cyl", count = FALSE, title = "hello", ylab = "YLAB Label")
}
