% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{make_latex_tables}
\alias{make_latex_tables}
\title{Title}
\usage{
make_latex_tables(file_name, table_title, table_reference_label, data,
  panel_caption, multi_column = NULL, footnote = "",
  sideways = FALSE)
}
\arguments{
\item{sideways}{}
}
\description{
Title
}
