% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_mean_median_mode_table_by_group}
\alias{make_mean_median_mode_table_by_group}
\title{Create a table showing the mean, median, and mode of a certain column}
\usage{
make_mean_median_mode_table_by_group(
  data,
  group_column,
  data_column,
  total_row = TRUE
)
}
\arguments{
\item{data}{A data.frame with the data you want to make the table from.}

\item{group_column}{A string with the name of the variable you are grouping by}

\item{data_column}{A string for the variable you want to get the  mean, median, and mode from,
Variable should be numeric.}

\item{total_row}{A boolean (default TRUE) for whether to include a row a the bottom for the
overall mean and standard deviation (i.e. not by group).}
}
\value{
A data.frame with the first column showing the category grouped by. Then one
column for the mean, one column for the median, and one column for the mode.
}
\description{
Create a table showing the mean, median, and mode of a certain column
}
\examples{
make_mean_median_mode_table_by_group(mtcars, "gear", "mpg")
}
