% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{make_state_abb}
\alias{make_state_abb}
\title{Returns abbreviations of state name input.}
\usage{
make_state_abb(state)
}
\arguments{
\item{state}{A vector of strings with the names of US states.}
}
\value{
A vector of strings with the abbreviations of the inputted state names.
}
\description{
Returns abbreviations of state name input.
}
\examples{
make_state_abb("california")
}
