% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_incidents.R
\name{philadelphia_incidents}
\alias{philadelphia_incidents}
\title{Philadelphia Incidents}
\usage{
philadelphia_incidents()
}
\value{
data.frame with all available Philly crime data
}
\description{
Philadelphia Incidents
}
\examples{
2 + 2
}

