% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theme.R
\name{scale_fill_crim}
\alias{scale_fill_crim}
\title{A set of colorblind friendly fill colors for graphs.}
\usage{
scale_fill_crim(...)
}
\arguments{
\item{...}{Arguments passed to discrete_scale()}
}
\value{
The ggplot graph with fills set.
}
\description{
A set of colorblind friendly fill colors for graphs.
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(x = cyl, fill = as.character(cyl))) +
 ggplot2::geom_bar() +
  scale_fill_crim()
}
