% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theme.R
\name{scale_linetype_crim}
\alias{scale_linetype_crim}
\title{A set of linetypes}
\usage{
scale_linetype_crim(...)
}
\arguments{
\item{...}{Arguments passed to discrete_scale()}
}
\value{
The ggplot graph with linetypes set.
}
\description{
A set of linetypes
}
\examples{
ggplot2::ggplot(mtcars, ggplot2::aes(x = mpg, y = hp, linetype = as.character(cyl))) +
  ggplot2::geom_line(size = 1) +
  scale_linetype_crim() +
  theme_crim()
}
