% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalizes the first letter of every word}
\usage{
simple_cap(word)
}
\arguments{
\item{word}{A string with words you want capitalized}
}
\value{
The origin string with the first letter of each word capitalized
}
\description{
Capitalizes the first letter of every word
}
