% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_incidents.R
\name{smart_date}
\alias{smart_date}
\title{Smartly apply lubridate}
\usage{
smart_date(dataset, column_name)
}
\arguments{
\item{dataset}{data.table of crime}

\item{column_name}{The string of the date column name}
}
\value{
A data.table
}
\description{
Smartly apply lubridate
}
\examples{
1* 3
}

