% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theme.R
\name{theme_crim}
\alias{theme_crim}
\title{A minimalistic theme designed for graphics in academic research}
\usage{
theme_crim()
}
\value{

}
\description{
A minimalistic theme designed for graphics in academic research
}
\examples{
ggplot2::ggplot(mtcars) +
ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg)) +
theme_crim()
}
