% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_codebooks.R
\name{UCR.LEOKANames}
\alias{UCR.LEOKANames}
\title{Uniform Crime Report Police Employee (LEOKA) Column Name Fixer}
\usage{
UCR.LEOKANames(UCR_dataset)
}
\arguments{
\item{UCR_dataset}{#' This is the data.frame containing data from the
UCR "Police Employee (LEOKA)" dataset
You may use any years from 1994-2014}
}
\value{
#' Returns the same data.frame as inputted but with column names corrected
as per the UCR "Police Employee (LEOKA)" codebook for years 1994-1997 or years
1998-2014
}
\description{
Uniform Crime Report Police Employee (LEOKA) Column Name Fixer
}
\examples{
# This is an example with 1994-1997 LEOKA column names
example <- data.frame(V1 = 1:10, V2 = 2:11, V8 = "example", V6 = 1995)
names(example)
example <- UCR.LEOKANames(example)
names(example)

# This is an example with 1998-2014 LEOKA column names
example <- data.frame(V1 = 1:10, V2 = 2:11, V8 = "example", V6 = 2000)
names(example)
example <- UCR.LEOKANames(example)
names(example)
}

