% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_codebooks.R
\name{UCR.OffenseNames}
\alias{UCR.OffenseNames}
\title{Uniform Crime Report Offenses and Clearances Column Name Fixer}
\usage{
UCR.OffenseNames(UCR_dataset)
}
\arguments{
\item{UCR_dataset}{This is the data.frame containing data from the
UCR "Offenses Known and Clearance by Arrest" dataset
You may use any years from 1998-2014}
}
\value{
Returns the same data.frame as inputted but with column names corrected
as per the UCR "Offenses Known and Clearance by Arrest" codebook for years
1998-2014
}
\description{
Uniform Crime Report Offenses and Clearances Column Name Fixer
}
\examples{
# This is an example data.frame with the same column names
# as the real UCR offenses and clearances, dataset.
example <- data.frame(V1 = 1:10, V2 = 2:11, V8 = "example")
names(example)
example <- UCR.OffenseNames(example)
names(example)

}

