% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_constant_reporters.R
\name{ucr_constant_reporter_oris}
\alias{ucr_constant_reporter_oris}
\title{Title}
\usage{
ucr_constant_reporter_oris(data, start_year, end_year,
  minimum_months_reported)
}
\arguments{
\item{minimum_months_reported}{}
}
\description{
Title
}
