% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_codebooks.R
\docType{data}
\name{UCR_Offenses_ColNames}
\alias{UCR_Offenses_ColNames}
\title{Column name and code for the UCR Offenses dataset}
\format{A data frame with 1448 rows and 2 variables:
\describe{
  \item{column_name}{The column name for the given column, according
  to the codebook.}
  \item{column_code}{The code for the given column, according to the codebook.
  This will be the column name of the inputted dataset.}
  ...
}}
\source{
\url{http://www.icpsr.umich.edu/icpsrweb/NACJD/studies/36391}
}
\usage{
UCR_Offenses_ColNames
}
\description{
A dataset containing the name and code corresponding to columns
in the Uniform Crime Report "Offenses Known and Clearance by Arrest"
dataset for years 1998-2014
}
\keyword{datasets}

