% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/police_incidents.R
\name{coordinate_splitter}
\alias{coordinate_splitter}
\title{Splits the coordinate column into a longitude and a latitude column}
\usage{
coordinate_splitter(dataset, column_name)
}
\arguments{
\item{dataset}{The crime dataset}

\item{column_name}{The name of the coordinates column (as a string)}
}
\value{
The dataset
}
\description{
Splits the coordinate column into a longitude and a latitude column
}
\examples{
2 + 2
}

