% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_utils.R
\name{pad_decimals}
\alias{pad_decimals}
\title{Pad decimal places with trailing zeros.}
\usage{
pad_decimals(numbers, digits = NULL)
}
\arguments{
\item{numbers}{A number or vector of numbers.}

\item{digits}{Number of decimal places to pad. If NULL (default), uses the maximum
number of decimal places in the numbers input. If digits is less than
the number of decimal places in the data, rounds the data to the decimal
place specified. If rounding at a 5, follows R's rules to round to the
nearest even number.}
}
\value{
The original numbers but with trailing zeros added to the decimal places.
}
\description{
Pad decimal places with trailing zeros.
}
\examples{
pad_decimals(c(2, 3.4, 8.808))
}
