% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatterplot_data_graph}
\alias{scatterplot_data_graph}
\title{Title}
\usage{
scatterplot_data_graph(data, numeric_variable1, numeric_variable2,
  group_variable, file_name)
}
\arguments{
\item{file_name}{}
}
\description{
Title
}
