% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{scatterplot_data_graph}
\alias{scatterplot_data_graph}
\title{Create a PDF with one scatterplot for each group in the data.}
\usage{
scatterplot_data_graph(
  data,
  numeric_variable1,
  numeric_variable2,
  group_variable,
  file_name
)
}
\arguments{
\item{data}{A data.frame with the data you want to graph.}

\item{numeric_variable1}{A string with the name of the first column with numeric data to graph.}

\item{numeric_variable2}{A string with the name of the second column with numeric data to graph.}

\item{group_variable}{A string with the name of the column with the grouping variable.}

\item{file_name}{A string with the name of the PDF to be made with one page for each graph.}
}
\value{
A PDF with one page per graph
}
\description{
Create a PDF with one scatterplot for each group in the data.
}
