% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucr_constant_reporters.R
\name{ucr_constant_reporter_oris}
\alias{ucr_constant_reporter_oris}
\title{Get ORIs that consistently report their data every year.}
\usage{
ucr_constant_reporter_oris(data, minimum_months_reported)
}
\arguments{
\item{data}{A data.frame with Uniform Crime Report (UCR) data. Requires at least
the ORI, year, and number_of_months_reported columns.}

\item{minimum_months_reported}{Integer indicating the minimum number of months requesting to keep in data.}
}
\value{
A vector with the ORIs that report the minimum number of months for every
year in the data.
}
\description{
Get ORIs that consistently report their data every year.
}
